+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Removed
 
 - Removed mocks package. [#9](https://github.com/go-nacelle/config/pull/9)
+- Removed `MustLoad` from `Config` interface. [#11](https://github.com/go-nacelle/config/pull/11)
 
 ### Changed